refactor: extract medication view renderers
